cleanups
- object initializers
- redundant type specifiers
- unnecessary public modifiers
- unnecessary null initialization for objects
- unnecessary parentheses
- unnecessary array type specifier
- unused variables
- nameof instead of hard-coded names
- unused namespace usings
- unneeded cast for enum values
- unnecessary boxing for nunit
- prefer async version of methods over sync

+++ src/Microsoft.Azure.ServiceBus/Primitives/MessagingEntityType.cs
@@ -10,6 +10,6 @@ namespace Microsoft.Azure.ServiceBus
         Subscriber = 2,
         Filter = 3,
         Namespace = 4,
-        Unknown = (int)0x7FFFFFFE,
+        Unknown = 0x7FFFFFFE,
     }
 }
